gst/va: move JPEG encoder test classes to lib
